database/requirements.txt
psycopg2-binary==2.9.9
psycopg[binary]>=3.1
pgvector>=0.2
python-dotenv>=1.0
